Refactor: eventHandler함수 재분리
