+++ src/Page/Main/index.js
@@ -1,7 +1,7 @@
 import ClickCard from '../../components/ClickCard/index.js';
 import HtmlElement from '../../core/HtmlElement.js';
 import { setInheritance } from '../../utils/manuplateDom.js';
-import { increaseTargetValue } from './eventHandler.js';
+import { handleClick } from './eventHandler.js';
 
 export default function Main($element) {
   HtmlElement.call(this, $element);
@@ -10,7 +10,9 @@ export default function Main($element) {
 setInheritance({ parent: HtmlElement, child: Main });
 
 Main.prototype.setTemplate = function () {
-  const { first, second, third, firth, mockArr } = this.store.state;
+  const {
+    state: { mockArr },
+  } = this.store;
   return mockArr
     ?.map((clickObj) => {
       const $clickcard = new ClickCard(this.$element, clickObj);
@@ -20,23 +22,6 @@ Main.prototype.setTemplate = function () {
     .join('');
 };
 
-Main.prototype.renderChild = function () {};
-
 Main.prototype.setEvent = function () {
-  const { handleClick } = this.eventHandler;
   this.$element.addEventListener('click', handleClick.bind(this));
-};
-
-Main.prototype.eventHandler = {
-  handleClick: function (event) {
-    const {
-      target: {
-        dataset: { clickType },
-      },
-    } = event;
-    if (!clickType) return;
-    const { mockArr } = this.store.state;
-    const newMockArr = increaseTargetValue(clickType, mockArr);
-    this.store.setState({ mockArr: newMockArr });
-  },
 };
